+++ src/globus_sdk/_testing/__init__.py
@@ -1,5 +1,5 @@
 from .helpers import get_last_request
-from .models import RegisteredResponse, ResponseSet
+from .models import RegisteredResponse, ResponseList, ResponseSet
 from .registry import (
     get_response_set,
     load_response,
@@ -10,6 +10,7 @@ from .registry import (
 __all__ = (
     "get_last_request",
     "ResponseSet",
+    "ResponseList",
     "RegisteredResponse",
     "load_response_set",
     "load_response",
