--- tests/functional/flows/test_create_flow.py
@@ -1,9 +1,0 @@
-from globus_sdk import FlowsClient
-from globus_sdk._testing import load_response
-
-
-def test_create_flow(flows_client: FlowsClient):
-    metadata = load_response(flows_client.create_flow).metadata
-
-    resp = flows_client.create_flow(**metadata["params"])
-    assert resp.data["title"] == "Multi Step Transfer"
